remove toolbar stuff
