+++ plugins/catalog/src/components/CatalogTable/CatalogTable.test.tsx
@@ -114,7 +114,7 @@ describe('CatalogTable component', () => {
         },
       },
     );
-    expect(screen.getByText(/Owned components \(3\)/)).toBeInTheDocument();
+    expect(screen.getByText(/Owned Components \(3\)/)).toBeInTheDocument();
     expect(screen.getByText(/component1/)).toBeInTheDocument();
     expect(screen.getByText(/component2/)).toBeInTheDocument();
     expect(screen.getByText(/component3/)).toBeInTheDocument();
