+++ plugins/catalog/src/components/CatalogTable/CatalogTableToolbar.tsx
@@ -14,15 +14,10 @@
  * limitations under the License.
  */
 import React from 'react';
-import {
-  EntitySearchBar,
-  useEntityList,
-} from '@backstage/plugin-catalog-react';
+import { EntitySearchBar } from '@backstage/plugin-catalog-react';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
-import { capitalize } from 'lodash';
-import pluralize from 'pluralize';
 
 /** @public */
 export type CatalogTableToolbarClassKey = 'root' | 'text';
@@ -45,28 +40,14 @@ const useToolbarStyles = makeStyles(
   { name: 'PluginCatalogTableToolbar' },
 );
 
-export function CatalogTableToolbar() {
+export function CatalogTableToolbar(props: {
+  title?: string | React.ReactElement<any>;
+}) {
   const styles = useToolbarStyles();
-
-  const { filters, totalItems } = useEntityList();
-
-  const currentKind = filters.kind?.label || '';
-  const currentType = filters.type?.value || '';
-  const currentCount = typeof totalItems === 'number' ? `(${totalItems})` : '';
-  const titlePreamble = capitalize(filters.user?.value ?? 'all');
-  const title = [
-    titlePreamble,
-    currentType,
-    pluralize(currentKind),
-    currentCount,
-  ]
-    .filter(s => s)
-    .join(' ');
-
   return (
     <Toolbar className={styles.root}>
       <Typography variant="h5" className={styles.text}>
-        {title}
+        {props.title}
       </Typography>
       <EntitySearchBar />
     </Toolbar>
