+++ plugins/catalog/src/components/CatalogTable/OffsetPaginatedCatalogTable.tsx
@@ -22,7 +22,6 @@ import {
   EntityTextFilter,
   useEntityList,
 } from '@backstage/plugin-catalog-react';
-import { CatalogTableToolbar } from './CatalogTableToolbar';
 
 /**
  * @internal
@@ -70,9 +69,6 @@ export function OffsetPaginatedCatalogTable(
       totalCount={totalItems}
       localization={{ pagination: { labelDisplayedRows: '' } }}
       isLoading={isLoading}
-      components={{
-        Toolbar: CatalogTableToolbar,
-      }}
     />
   );
 }
